# static/openapi.yaml
openapi: "3.0.3"
info:
  title: Ofir Assulin Performance Marketing API
  version: "1.0.0"
  description: |
    API for the autonomous AI-powered performance marketing system for Ofir Assulin Interior Design.
    Manages leads, campaigns, proposals, landing pages, brand assets, integrations, A/B tests, and analytics.
servers:
  - url: http://localhost:5000
    description: Local development
paths:
...
  /api/leads:
    get:
      summary: List leads
      tags: [Leads]
      parameters:
        - {name: status, in: query, schema: {type: string}}
        - {name: source, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, default: 1}}
        - {name: per_page, in: query, schema: {type: integer, default: 50}}
      responses:
        "200":
          description: Paginated list of leads
    post:
      summary: Create a new lead
      tags: [Leads]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, phone]
              properties:
                name: {type: string}
                phone: {type: string}
                email: {type: string}
                source: {type: string, default: organic}
                campaign: {type: string}
                location: {type: string}
                budget: {type: string}
                project_type: {type: string}
      responses:
        "201":
          description: Lead created and scored
  /api/leads/{lead_id}/status:
    put:
      summary: Update lead status
      tags: [Leads]
      parameters:
        - {name: lead_id, in: path, required: true, schema: {type: string}}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: {type: string}
                notes: {type: string}
      responses:
        "200":
          description: Status updated
...